refactor: streamline HeatMap component and remove unused imports
- Simplified HeatMap component by directly calculating the center from measurements instead of using a utility function.
- Updated CircleMarker to use a new position variable for better clarity.
- Removed unused setBounds prop from DrawLayer and FilteringMapModal components to enhance code cleanliness.
- Cleaned up imports in FilteringMapModal for improved organization.
