+++ src/app/Home/_components/CampaignFilterToolbar/_components/FilteringMapModal/FilteringMapModal.tsx
@@ -58,7 +58,6 @@ const FilteringMapModal: React.FC<FilteringModalProps> = ({
             onCreate={handleCreate}
             onChange={handleChange}
             bounds={bounds}
-            setBounds={setBounds}
           />
         </MapContainer>
       </div>
